requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.27.0

# LangChain + LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0

# RAG - ChromaDB
chromadb>=0.5.0
sentence-transformers>=2.7.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0

# Database
sqlalchemy>=2.0.0

# MLflow Tracking
mlflow>=2.14.0

# Telegram Bot
python-telegram-bot>=21.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Scheduling
apscheduler>=3.10.0

# Utilities
tenacity>=8.2.0
pydantic>=2.7.0
